Refactored code design of all tests
